requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.6.1
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
streamlit==1.28.0
streamlit-chat==0.1.1 